+++ IAParams.h
@@ -18,7 +18,7 @@ namespace params
   // Power for the difference between two path points value for similarity (this difference is always <= 1)
   extern float similarityTolerance;
   // Propagation of paths radius
-  extern float propagationRadius;
+  extern int propagationRadius;
 
   // ## Path linking resource and city detection - TILES VERSION
   // Percentage of the InfluenceMap of the resource/city needed to be considered as linking
